robot/requirements.txt
python-binance==1.0.19
requests==2.31.0
pandas==2.2.0
numpy==1.26.0
plotly==5.17.0
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
schedule==1.2.0
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
alembic==1.13.1
jinja2-cli==0.7.0
cryptography==42.0.0
